requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
coverage==7.6.1
cryptography==43.0.1
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==4.2.15
django-allauth==0.61.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
environs==11.0.0
exceptiongroup==1.2.2
factory_boy==3.3.1
Faker==30.1.0
flake8==7.1.1
flake8-isort==6.1.1
idna==3.8
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jwt==1.3.1
marshmallow==3.22.0
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
pytest-factoryboy==2.7.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.6
